+++ net-reporter/src/continuous_aggregate/bandwidth_per_endpoint.rs
@@ -22,9 +22,10 @@ impl ContinuousAggregate for BandwidthPerEndpointAggregate {
                     agent_id,
                     src_addr,
                     dst_addr,
-                    (binary_data->'l1'->'frame'->>'frame.len')::integer as packet_length
+                    (binary_data->'l1'->'frame'->>'frame.len')::integer as packet_length,
+                    binary_data->'l1'->'frame'->>'frame.protocols' as protocols
                 FROM captured_traffic
-                GROUP BY bucket, group_id, agent_id, src_addr, dst_addr, packet_length;
+                GROUP BY bucket, group_id, agent_id, src_addr, dst_addr, packet_length, protocols;
             ",
             Self::get_name()
         );